# openapi.yml
openapi: 3.0.0
info:
  title: TofunoteAPI
  description: ユーザーの日記を管理するためのRESTful API
  version: 1.0.0

servers:
  - url: http://localhost:8080/api
    description: ローカル開発環境
  - url: https://api.tofunote.takoscreamo.com/api
    description: 本番環境

security:
  - BearerAuth: []

paths:
  /me/diaries:
    get:
      summary: 日記一覧取得
      description: 現在のユーザーの日記一覧を取得します
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Diary'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: 日記作成
      description: 現在のユーザーの新しい日記を作成します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDiaryDTO'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Diary'
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 複合ユニークキー制約違反（同じユーザーの同じ日付の日記が既に存在）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me/diaries/range:
    get:
      summary: 期間指定日記取得
      description: 現在のユーザーの指定された期間の日記を取得します
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: 開始日（YYYY-MM-DD形式）
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: 終了日（YYYY-MM-DD形式）
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Diary'
        '400':
          description: リクエストが不正（パラメータ不足）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me/diaries/{date}:
    get:
      summary: 日記取得
      description: 現在のユーザーの指定された日付の日記を取得します
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: 日付（YYYY-MM-DD形式）
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Diary'
        '400':
          description: リクエストが不正（無効な日付）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 指定された日記が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: 日記更新
      description: 現在のユーザーの指定された日付の日記を更新します
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: 日付（YYYY-MM-DD形式）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDiaryDTO'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Diary'
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 指定された日記が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: 日記削除
      description: 現在のユーザーの指定された日付の日記を削除します
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: 日付（YYYY-MM-DD形式）
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: 成功メッセージ
                    example: "日記が正常に削除されました"
        '400':
          description: リクエストが不正（無効な日付）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 指定された日記が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me:
    get:
      summary: ユーザー情報取得
      description: 現在のユーザーの情報を取得します
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  nickname:
                    type: string
        '401':
          description: 認証情報が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: ユーザー情報部分更新
      description: 現在のユーザーの情報を部分的に更新します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: ユーザー情報を更新しました
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      nickname:
                        type: string
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証情報が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: ユーザー削除
      description: 現在のユーザーのアカウントを削除します
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: ユーザーアカウントを削除しました
        '401':
          description: 認証情報が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: ユーザーが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /guest-login:
    post:
      summary: ゲストログイン
      description: サーバー側でUUIDとリフレッシュトークンを生成し、ゲストユーザーとしてJWTトークンを発行する
      responses:
        '200':
          description: ゲストトークン発行成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refresh_token:
                    type: string
                  id:
                    type: string
                    format: uuid
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /refresh-token:
    post:
      summary: JWTリフレッシュ
      description: リフレッシュトークンを使って新しいJWTを発行する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: リフレッシュトークン
              required:
                - refresh_token
      responses:
        '200':
          description: 新しいJWT発行成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: リフレッシュトークン無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Diary:
      type: object
      properties:
        id:
          type: integer
          description: 日記ID
        user_id:
          type: integer
          description: ユーザーID
        date:
          type: string
          format: date
          description: 日付（YYYY-MM-DD形式）
        mental:
          type: integer
          minimum: 1
          maximum: 10
          description: メンタルスコア（1〜10の整数）
        diary:
          type: string
          description: 日記の内容
      required:
        - id
        - user_id
        - date
        - mental
        - diary

    Mental:
      type: integer
      minimum: 1
      maximum: 10
      description: メンタルスコア（1〜10の整数）

    CreateDiaryDTO:
      type: object
      properties:
        date:
          type: string
          format: date
          description: 日付（YYYY-MM-DD形式）
        mental:
          type: integer
          minimum: 1
          maximum: 10
          description: メンタルスコア（1〜10の整数）
        diary:
          type: string
          description: 日記の内容
      required:
        - date
        - mental
        - diary

    UpdateDiaryDTO:
      type: object
      properties:
        mental:
          type: integer
          minimum: 1
          maximum: 10
          description: メンタルスコア（1〜10の整数）
        diary:
          type: string
          description: 日記の内容
      required:
        - mental
        - diary

    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
      required:
        - error